+++ packages/excalidraw/element/newElement.ts
@@ -136,13 +136,9 @@ export const newElement = (
 export const newEmbeddableElement = (
   opts: {
     type: "embeddable";
-    validated: ExcalidrawEmbeddableElement["validated"];
   } & ElementConstructorOpts,
 ): NonDeleted<ExcalidrawEmbeddableElement> => {
-  return {
-    ..._newElementBase<ExcalidrawEmbeddableElement>("embeddable", opts),
-    validated: opts.validated,
-  };
+  return _newElementBase<ExcalidrawEmbeddableElement>("embeddable", opts);
 };
 
 export const newIframeElement = (